backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-genai>=1.0.0
python-multipart==0.0.9
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.7.4
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
apscheduler==3.10.4
bcrypt==4.1.3
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1
groq==0.9.0
celery[redis]==5.3.6
redis==5.0.3
openai==1.30.1
sendgrid==6.11.0
dnspython==2.6.1
httpx==0.27.0
razorpay==1.4.1
